% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmNormalization.R
\name{cpmNormalization}
\alias{cpmNormalization}
\title{cpmNormalization}
\usage{
cpmNormalization(X)
}
\arguments{
\item{X}{...}
}
\value{
...
}
\description{
Perfrom counts per million (CPM) normalization
}
