% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installPyDependencies.R
\name{installPyDependencies}
\alias{installPyDependencies}
\title{Install \code{Phyton} dependencies}
\usage{
installPyDependencies()
}
\description{
Install the \code{numpy} and \code{scipy} \code{Phyton} packages using \code{reticulate}.
}
