% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifoldAlignment.R
\name{manifoldAlignment}
\alias{manifoldAlignment}
\title{manifoldAlignment}
\usage{
manifoldAlignment(X, Y, d = 3)
}
\arguments{
\item{X}{...}

\item{Y}{...}

\item{d}{...}
}
\value{
...
}
\description{
...
}
\details{
...
}
\references{
...
}
