% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCrTdMa.R
\name{PCrTdMa}
\alias{PCrTdMa}
\title{PCrTdMa}
\usage{
PCrTdMa(X, Y, nNet = 10, nCells = 1000, nComp = 3,
  symmetric = FALSE, scaleScores = TRUE, q = 0.05)
}
\arguments{
\item{X}{...}

\item{Y}{...}

\item{nNet}{...}

\item{nCells}{...}

\item{nComp}{...}

\item{symmetric}{...}

\item{scaleScores}{...}

\item{q}{...}
}
\value{
...
}
\description{
...
}
\references{
...
}
