% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCoexpression.R
\name{dCoexpression}
\alias{dCoexpression}
\title{dCoexpression}
\usage{
dCoexpression(manifoldOutput, nGenes, geneList, q = 0.05)
}
\arguments{
\item{manifoldOutput}{...}

\item{nGenes}{...}

\item{geneList}{...}

\item{q}{...}
}
\value{
...
}
\description{
...
}
\references{
...
}
