% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNetworks.R
\name{makeNetworks}
\alias{makeNetworks}
\title{makeNetworks}
\usage{
makeNetworks(X, nNet = 10, nCells = 1000, nComp = 3,
  scaleScores = TRUE, symmetric = FALSE, q = 0.95)
}
\arguments{
\item{X}{...}

\item{nNet}{...}

\item{nCells}{...}

\item{nComp}{...}

\item{scaleScores}{...}

\item{symmetric}{...}

\item{q}{...}
}
\value{
...
}
\description{
...
}
\references{
...
}
