% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifoldAlignment.R
\name{manifoldAlignment}
\alias{manifoldAlignment}
\title{manifoldAlignment}
\usage{
manifoldAlignment(X, Y, d = 30)
}
\arguments{
\item{X}{A gene regulatory matrix.}

\item{Y}{A gene regulatory matrix.}

\item{d}{The dimension of the low-dimensional feature space.}
}
\value{
A low-dimensional projection for two gene regulatory matrices.
}
\description{
Build comparable low-dimensional features for the genes from two conditions.
}
\details{
...
}
\references{
...
}
