% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifoldAlignment.R
\name{manifoldAlignment}
\alias{manifoldAlignment}
\title{manifoldAlignment}
\usage{
manifoldAlignment(X, Y, wX, wY, d = 3, method = "nonLinear")
}
\arguments{
\item{X}{...}

\item{Y}{...}

\item{wX}{...}

\item{wY}{...}

\item{d}{...}

\item{method}{...}
}
\value{
...
}
\description{
...
}
\details{
...
}
\references{
...
}
