% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcNet.R
\name{pcNet}
\alias{pcNet}
\title{pcNet}
\usage{
pcNet(X, nCom = 3, scaleScores = TRUE, symmetric = FALSE, q = 0)
}
\arguments{
\item{X}{...}

\item{nCom}{...}

\item{scaleScores}{...}

\item{symmetric}{...}

\item{q}{...}
}
\value{
...
}
\description{
...
}
\details{
...
}
\examples{
# Generating input matrix
set.seed(1)
inputMatrix <- matrix(data = rnbinom(n = 1000, size = 10, prob = .9), nrow = 20)
rownames(inputMatrix) <- paste0('Gene_', seq_len(nrow(inputMatrix)))

# Getting the adjacency matrix
pcNet(inputMatrix)
# Getting the adjacency matrix filtered to the top 1\%
pcNet(inputMatrix, q = 0.99)
# Getting the adjacency matrix filtered to the top 5\%
pcNet(inputMatrix, q = 0.95)
# Getting the adjacency matrix filtered to the top 10\%
pcNet(inputMatrix, q = 0.90)
}
\references{
...
}
