% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcNet.R
\name{pcNet}
\alias{pcNet}
\title{pcNet}
\usage{
pcNet(X, nComp = 3, scaleScores = TRUE, symmetric = FALSE, q = 0)
}
\arguments{
\item{X}{A filtered and normlized gene expression matrix with cells as columns and genes as rows.}

\item{nComp}{An integer value. The number of principal components in PCA to generate the networks. Should be greater than 2 and lower than the total number of genes.}

\item{scaleScores}{A boolean value (TRUE/FALSE), if TRUE, the weights will be normalized such that the maximum absolute value is 1.}

\item{symmetric}{A boolean value (TRUE/FALSE), if TRUE, the weights matrix returned will be symmetric.}

\item{q}{A decimal value between 0 and 1. Represent the cut-off threshold of top q\% relationships to be returned.}
}
\value{
A gene regulatory network in dgCMatrix format.
}
\description{
Generate a gene regulatory network based on principal components regression.
}
\details{
...
}
\references{
...
}
