% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scQC.R
\name{scQC}
\alias{scQC}
\title{scQC}
\usage{
scQC(X, minLibSize = 1000, removeOutlierCells = TRUE, minPCT = 0.05,
  maxMTratio = 0.1)
}
\arguments{
\item{X}{...}

\item{minLibSize}{...}

\item{removeOutlierCells}{...}

\item{minPCT}{...}

\item{maxMTratio}{...}
}
\value{
...
}
\description{
...
}
\references{
...
}
