% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTenifoldNet.R
\name{scTenifoldNet}
\alias{scTenifoldNet}
\title{scTenifoldNet}
\usage{
scTenifoldNet(X, Y, nNet = 10, nCells = 500, nComp = 3,
  symmetric = FALSE, scaleScores = TRUE, q = 0.05)
}
\arguments{
\item{X}{...}

\item{Y}{...}

\item{nNet}{...}

\item{nCells}{...}

\item{nComp}{...}

\item{symmetric}{...}

\item{scaleScores}{...}

\item{q}{...}
}
\value{
...
}
\description{
...
}
\references{
...
}
