% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorDecomposition.R
\name{tensorDecomposition}
\alias{tensorDecomposition}
\title{tensorDecomposition}
\usage{
tensorDecomposition(xList, yList, d = 5)
}
\arguments{
\item{xList}{...}

\item{yList}{...}

\item{d}{...}
}
\value{
...
}
\description{
...
}
\details{
...
}
\references{
...
}
